feat: 新增用户统计页面，并作为默认路由
